feat: keep-starknet-strange#947: base lru cache
This commit introduces a base LRU cache that will be used for the RPC layer. It is inspired by frontier's code base:
https://github.com/paritytech/frontier/blob/194f6bb06152402ba44b340c3d401ae6e0670d96/client/rpc/src/eth/cache/mod.rs#L76

Extending it, our approach proposes two types of limiter:
- Values size limiter: A limiter that will base its cache management on the total size of all values stored in the cache
- Allocated size limiter: A limiter that will base its cache management on the allocated memory for the cache

This allows for either an approximate or a precise memory management, which could prove useful in our P2P context.